# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Football Training Board API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /training-sessions:
    get:
      operationId: listTrainingSessions
      tags: [training]
      summary: List all training sessions in the club
      parameters:
        - name: teamId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of training sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TrainingSession"
    post:
      operationId: createTrainingSession
      tags: [training]
      summary: Create a new training session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTrainingSessionRequest"
      responses:
        "201":
          description: Training session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrainingSession"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /training-sessions/{id}:
    get:
      operationId: getTrainingSession
      tags: [training]
      summary: Get a training session by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Training session details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrainingSession"
        "404":
          description: Training session not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateTrainingSession
      tags: [training]
      summary: Update a training session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTrainingSessionRequest"
      responses:
        "200":
          description: Updated training session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrainingSession"
    delete:
      operationId: deleteTrainingSession
      tags: [training]
      summary: Delete a training session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Training session deleted
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    TrainingSession:
      type: object
      properties:
        id:
          type: integer
        clubId:
          type: integer
        teamId:
          type: ["integer", "null"]
        teamName:
          type: ["string", "null"]
        title:
          type: string
        description:
          type: ["string", "null"]
        scheduledAt:
          type: string
          format: date-time
        durationMinutes:
          type: ["integer", "null"]
        location:
          type: ["string", "null"]
        status:
          type: string
        objectives:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - clubId
        - title
        - scheduledAt
        - status
        - createdAt
    CreateTrainingSessionRequest:
      type: object
      properties:
        teamId:
          type: ["integer", "null"]
        title:
          type: string
        description:
          type: ["string", "null"]
        scheduledAt:
          type: string
          format: date-time
        durationMinutes:
          type: ["integer", "null"]
        location:
          type: ["string", "null"]
        status:
          type: string
        objectives:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
      required:
        - title
        - scheduledAt
        - status
    UpdateTrainingSessionRequest:
      type: object
      properties:
        teamId:
          type: ["integer", "null"]
        title:
          type: string
        description:
          type: ["string", "null"]
        scheduledAt:
          type: string
          format: date-time
        durationMinutes:
          type: ["integer", "null"]
        location:
          type: ["string", "null"]
        status:
          type: string
        objectives:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
...